// src/data.d.ts
export interface DatabaseType {
  providers: Providers;
  locations: Locations;
  routes: Routes;
  worlds: Worlds;
  routesByLocation: RoutesByLocation;
}

export interface Providers {
  [key: string]: Provider;
}

export interface Provider {
  name: string;
  numberPrefix: string | null;
  description: string | null;
  logo: string | null;
  color: Color;
  ownerPlayer: string | null;
  alias: Alias[] | null;
  autoGenerated: boolean;
}

export interface Alias {
  use: string;
  numberRange: NumberRange;
}

export interface NumberRange {
  start: number;
  end: number;
}

export interface Color {
  light: string;
  dark: string;
}

export interface Locations {
  [key: string]: Location;
}

export interface Location {
  name: string;
  shortName: string;
  MRT_TRANSIT_NAME: string | null;
  IATA: string | null;
  description: string | null;
  location: Coordinates | null;
  ownerPlayer: string | null;
  world: string;
  enabled: boolean;
  autoGenerated: boolean;
  type: "City" | "Airport" | "MRT Station" | "Other";
  isSpawnWarp: boolean;
  uniqueId?: string;
}

export interface Coordinates {
  x: number;
  y: number | null;
  z: number;
}

export interface Routes {
  [key: string]: Route;
}

export interface Route {
  name: string | null;
  description: string | null;
  locations: RouteLocations;
  type: "flight" | "seaplane" | "heli" | "MRT" | "walk" | "spawnWarp";
  autoGenerated: boolean;
  provider: string;
}

export interface RouteLocations {
  [key: string]: string;
}

export interface Worlds {
  [key: string]: World;
}

export interface World {
  name: string;
}

export interface RoutesByLocation {
  [key: string]: RouteByLocation;
}

export interface RouteByLocation {
  autoGenerated: boolean;
  routes: string[];
}